requirements.txt
flask
flask_jwt_extended
flask_smorest
flask_sqlalchemy
marshmallow
pre-commit
psycopg2-binary
python-dotenv
supabase

